#146 — Client: rename CombatResult* → ContestResult* (no behavior change)
Phase 2 of #146. Pure rename — combat is treated as a strength-stat contest
under one UI umbrella, so the type/file/component/state-export names all
move to the contest vocabulary. No code paths change; tests reflect the new
import paths only.

Renames:
- src/lib/combatResult.ts → contestResult.ts (CombatPairResolved engine
  event type kept as-is; it refers to the engine `combat_pair_resolved`
  event name, which is not part of this rename).
- src/lib/__tests__/combatResult.test.ts → contestResult.test.ts
- src/components/CombatResultPopup.svelte → ContestResultPopup.svelte

Identifier renames in gameStore + popup + App + Modal comment:
- type CombatResult → ContestResult
- type CombatOutcome → ContestOutcome
- _combatResult / getCombatResult / dismissCombat →
  _contestResult / getContestResult / dismissContest
