+++ clients/web/src/lib/contestResult.ts
